Refactor into a single python distribution
There's no sense distributing relengapi-docs separately from relengapi.

Most of this just moves ./docs/* to ./base/*.  The rest fixes up
the lint options and tests so that validation still passes.

We'll keep the base/ subdirectory for a while, to allow other relengapi
tools to "incubate" in this repository while it's still in flux.
Eventually all of this will move to the top level.

Fixes #59.