Adds the ability to clobber specific slaves
The clobber endpoint now supports passing in a specific slave, when a slave is specified
only lastclobber queries with that slave will have access to its clobbertime. Otherwise,
when no slave is specified, a NULL value is passed to the clobber_times table and any
slave will have access to its data. That is, NULL acts as a wildcard for slaves.