Bug 1164878 - Allow users without an authenticated_email field to clobber
The TokenUser is not anonymous, but also has no e-mail address associated with it.
When this combination occurs, a who value of 'automation' should be supplied, instead
of outright failing.

While not exact, the term automation gives us a way to categorize clobbers that seemed
to be authenticated by a permanent token (as opposed to a user token).