Combine all tree modifictions into one method
The PATCH is always applied to /trees, with the specific tree(s)
affected specified in the body.  The various fields of a tree update
become optional.  This makes the business rules a little complicated,
but maps well to the desired UI, which entails submitting a bulk update
to the whole list of trees.