Update logic for updating time to live
You can only update time to live if the file already has an existing
expiration date, or is a new file with a time to live value. You may
also update a file to make it permanent, which you can no longer update
the time to live value.

This patch also brings in coverage tests, as well as using
relengapi.lib.time instead of datetime.datetime/