releng_frontend/trychooser: Remove the dead platforms android-api-15-frontend, sm-warnaserr, and sm-generational (#285)
android-api-15-frontend was renamed to android-test a year ago[1], and the fact that nobody has made even the tiniest peep about not knowing how to trigger it is a sign it's not needed; if someone wants to run it, they can add it with whatever name it has then (and I hope that name is better than A Build Named Test).

sm-generational was removed a year ago[2], a year after forgetting that it was only left around[3] for the period between landing an in-tree patch and a buildbot patch.

Bug archaeology of the death of sm-warnaserr is... not recommended. There are things like a three-year period when we had two different configure args for enabling warnings-as-errors, both of which actually disabled warnings-as-errors. It doesn't exist, attempting to run it now causes the taskcluster decision-task to fail, that's enough, it should be gone.


[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1260874
[2] https://bugzilla.mozilla.org/show_bug.cgi?id=1272720
[3] https://bugzilla.mozilla.org/show_bug.cgi?id=1136309