releng_frontend/trychooser: Fix inaccurate info in reftest/crashtest labels, and give up on listing chunks (#287)
If there's one thing that's true about trychooser/try syntax, it's that there's no one true thing about it.

Despite the fact that I'll shortly have a PR that adds more individual chunks for mochitests, it doesn't make sense to maintain a list of chunks for reftests: they run in a bewildering variety of chunks (8 on Linux, 1 on Mac and Windows opt, 2 on Windows debug, and only the 2 e10s on Mac debug, and insane 16 and 48 chunks on Android, which I'm not yet prepared to touch), and unlike mochitests where it might be reasonable to run only M10 because you're working on toolkit/components/alerts/ on Linux64 and don't need to run totally unrelated browser/components/feeds/ tests in M1, with reftests you're generally working on... layout stuff, and you need to run the reftest chunks that deal with layout stuff, which is all of them, all the more so because on Mac you have no choice but to run them all, and on Windows no choice but to run at least half of them. Even if we did list reftest-1 through reftest-10, there would be absolutely no sane way to label them, since accurate labels would be things like "reftest-1 (linux and windows only, but totally utterly different things on linux and on windows and for mac you have to pick just 'reftest' anyway because reftest-1 doesn't run reftest when there's only one, so why are you here at all?)"

Removing the info span on crashtest-e10s is just a driveby, we long since started running it on Mac and Win8.