please_cli: Change volume use to enable Mac use
Macs appear to have an issue with their host:volume mounts using osxfs, and extended attribute permissions, meaning the nix store becomes unreadable during setup if using `cp` to populate it.

Change to use a docker volume rather than a host directory.
If the named volume is empty, and the mount point is not, the mount point's contents are copied into the new volume before the mount occurs. This gives us the same behaviour as the host directory version, but using a named volume. It does make it less easy to modify `/nix/store` when outside a container, but does enable people to run macs.