shipit_code_coverage: Fix bug that prevented us from actually waiting for coveralls to injest data (#355)
The check_coveralls_job function is returning True when the build is injested and False when it isn't. The wait_until function was checking for the return value to be not None. This meant that, even when the build was not injested and check_coveralls_job returned False, wait_until was completing.

With this change, we are actually waiting coveralls to injest the build before generating the coverage by dir report.
I've removed the code which raised an exception when Coveralls fails to injest a report, as it's too likely to happen for now.