shipit-workflow: Use decision task for in action context's `taskGroupId`.
The action specification[1] documents this as being set to the decision
task id of the task that defined the actions.

There is also discussion in scriptworker docs[2] suggesting that it is set
differently for release promotion tasks, so that the created tasks get
created in a new graph. However, it turns out that release-promotion always
creates tasks in a task group named after the action tasks[3][4][5].

[1] https://docs.taskcluster.net/docs/manual/using/actions/spec#json-e-context
[2] https://github.com/mozilla-releng/scriptworker/blob/77bb2f2adfd08ddb22e6cec4d8bc377a58222e87/scriptworker/cot/verify.py#L1240-L1245
[3] https://searchfox.org/mozilla-central/rev/f822a0b61631cbb38901569e69b4967176314aa8/taskcluster/taskgraph/actions/release_promotion.py#313
[4] https://searchfox.org/mozilla-central/rev/f822a0b61631cbb38901569e69b4967176314aa8/taskcluster/taskgraph/decision.py#173
[5] https://searchfox.org/mozilla-central/rev/f822a0b61631cbb38901569e69b4967176314aa8/taskcluster/taskgraph/create.py#41-47