codecoverage/bot: Remove now useless 'changeset' parameter from phabricator module
It is not needed anymore since we can now always use json-automationrelevance, so even for
the try repository we will retrieve all the changesets we need in the phabricator module.
Before, in the phabricator module we were using json-pushes, which does not return past
changesets that are relevant for the try push but are not in the latest push (e.g. patches
on which your latest patch sent to try is based).