Change clang-format's command message to show, not apply
`./mach clang-format -p path/to/file.cpp` applies changes silently to the working tree. `COMMENT_FAILURE` says "You can run this analysis locally with:" which to me suggests a "show, don't do" command.

By putting in `-s` that instructs `clang-format` to perform the analysis and show the results instead of silently applying them.

To be clear this doesn't change how clang-format is run against patches, just how it's suggested to be run in the helpful reviewbot message.